requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
dotenv==1.0.0

pdfplumber==0.10.3

sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0+cpu --index-url https://download.pytorch.org/whl/cpu
transformers==4.35.0
tokenizers==0.15.0
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.24.4

# Utilities
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
aiofiles==23.2.1

# Production Server
gunicorn==21.2.0
